Fix predicated ternary cleanup and regenerate goldens

+++ test_cu/golden_lifted/loop_constant.pseudo.c
@@ -7,17 +7,17 @@ BB1 {
   R2.0 = threadIdx.x;
   UR6.0 = ConstMem(0, 280);
   UR7.0 = ConstMem(0, 284);
-  P1.0 = R2.0 > 255;
+  P1.0 = (int32_t)(R2.0) > (int32_t)(255);
   P0.1 = R2.0 != 0;
 }
 // Condition from BB1
-if (!(R2.0 > 255)) {
+if (!((int32_t)(R2.0) > (int32_t)(255))) {
   BB2 {
     shmem_u8[R2.1] = R2.1;
     R2.2 = R2.1 + ConstMem(0, 0);
   }
   // Loop header BB2
-  while (!(R2.2 >= 256)) {
+  while (!((int32_t)(R2.2) >= (int32_t)(256))) {
     BB2 {
       shmem_u8[R2.1] = R2.1;
       R2.2 = R2.1 + ConstMem(0, 0);
